Safari-611.1.10/LayoutTests/fast/forms/select/listbox-disabled-scroll-no-onchange-expected.txt

Test that scrolling the listbox does not fire change event.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS changeEventCount is 0
PASS successfullyParsed is true

TEST COMPLETE
